requirements.txt
Django==3.1.3
djangorestframework==3.12.2
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@bd15a17163d25bcf919e31a97d87eefba531bcef
omegaconf==2.0.5
parallel_wavegan==0.4.8
scipy==1.5.4
torch==1.6.0
megatron-lm==1.1.5
braceexpand
webdataset
frozendict
inflect
unidecode
g2p_en
pandas
torch-stft
editdistance
espnet==0.9.3
